Move sidebar tests to Sidebar.spec.tsx
