+++ packages/test-support/src/keyring/addresses.ts
@@ -3,5 +3,9 @@
 
 export const alice = '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY';
 export const bob = '5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty';
+export const charlie = '5DAAnrj7VHTznn2AWBemMuyBwZWs6FNFjdyVXUeYum3PTXFy';
 export const ferdie = '5CiPPseXPECbkjWCa6MnjNokrgYjMqmKndv2rSnekmSK2DjL';
+
+export const defaultAddresses = [alice, bob, charlie];
+
 export const defaultMembers = { isMember: true, members: [alice, bob, ferdie] };
