+++ packages/test-support/src/keyring/index.ts
@@ -3,3 +3,4 @@
 
 export * from './MemoryStore';
 export * from './signers';
+export * from './addresses';
